pyproject.toml
[project]
name = "gdrive-cli"
version = "0.1.0"
description = "Google Drive CLI for fetching and searching docs, sheets, and slides"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.scripts]
gdrive-cli = "gdrive_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gdrive_cli"]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
